Divide prop tests to different modules
